Revert "Port `BeforeTypeCast` to Active Model"
